# perf(request): avoid USVString in RequestInfo