--- a/lib/fetch/request.js
+++ b/lib/fetch/request.js
@@ -50,7 +50,7 @@ class Request {
 
     webidl.argumentLengthCheck(arguments, 1, { header: 'Request constructor' })
 
-    input = webidl.converters.RequestInfo(input)
+    input = webidl.converters.RequestInfo_DOMString(input)
     init = webidl.converters.RequestInit(init)
 
     // https://html.spec.whatwg.org/multipage/webappapis.html#environment-settings-object
@@ -897,6 +897,19 @@ webidl.converters.RequestInfo = function (V) {
   }
 
   return webidl.converters.USVString(V)
+}
+
+// DOMString is used because the value is converted to a USVString in `new URL()`.
+webidl.converters.RequestInfo_DOMString = function (V) {
+  if (typeof V === 'string') {
+    return webidl.converters.DOMString(V)
+  }
+
+  if (V instanceof Request) {
+    return webidl.converters.Request(V)
+  }
+
+  return webidl.converters.DOMString(V)
 }
 
 webidl.converters.AbortSignal = webidl.interfaceConverter(